Commit for 1.2 Version

+++ Es.Data/Models/MemberUsersRolesModel.cs
@@ -3,9 +3,9 @@
     public class MemberUsersRolesModel
     {
         public System.Guid Id { get; set; }
-        public long MemberRoleId { get; set; }
-        public long EsUserId { get; set; }
-        public long MemberId { get; set; }
+        public int MemberRoleId { get; set; }
+        public int EsUserId { get; set; }
+        public int MemberId { get; set; }
 
         public virtual EsMemberModel EsMembers { get; set; }
         public virtual EsUserModel EsUsers { get; set; }
